feat(cli): refactor user creation logic into separate module and improve validation checks
