+++ packages/studiocms/src/cli/cmds/users/steps/libsql/createUsers.ts
@@ -1,13 +1,13 @@
 import { z } from 'astro/zod';
 import color from 'chalk';
 import dotenv from 'dotenv';
-import checkIfUnsafe from '../../../../lib/auth/utils/unsafeCheck.js';
-import type { Context } from '../../../lib/context.js';
-import { tsPermissions, tsUsers, useLibSQLDb } from '../../../lib/useLibSQLDb.js';
-import { StudioCMSColorwayError, StudioCMSColorwayInfo } from '../../../lib/utils.js';
-import { createUserAvatar } from './utils/avatar.js';
-import { checkRequiredEnvVars } from './utils/checkRequiredEnvVars.js';
-import { hashPassword } from './utils/password.js';
+import checkIfUnsafe from '../../../../../lib/auth/utils/unsafeCheck.js';
+import type { Context } from '../../../../lib/context.js';
+import { tsPermissions, tsUsers, useLibSQLDb } from '../../../../lib/useLibSQLDb.js';
+import { StudioCMSColorwayError, StudioCMSColorwayInfo } from '../../../../lib/utils.js';
+import { createUserAvatar } from '../utils/avatar.js';
+import { checkRequiredEnvVars } from '../utils/checkRequiredEnvVars.js';
+import { hashPassword } from '../utils/password.js';
 
 dotenv.config();
